backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
pypdf
python-docx
python-multipart
pandas
openpyxl
